requirements.txt
PySocks==1.7.1
allure-pytest==2.13.2
allure-python-commons==2.13.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
execnet==2.0.2
h11==0.14.0
idna==3.6
iniconfig==2.0.0
outcome==1.3.0.post0
packaging==23.2
pip==23.2.1
pluggy==1.4.0
pycparser==2.21
pytest==8.0.1
pytest-xdist==3.5.0
requests==2.31.0
selenium==4.18.1
setuptools==65.5.0
sniffio==1.3.0
sortedcontainers==2.4.0
trio==0.24.0
trio-websocket==0.11.1
typing-extensions==4.9.0
urllib3==2.2.1
wsproto==1.2.0